refactor: move `charmcraft.yaml` to top level
This is a simplification that pulls the charmcraft
build file to the top level of the repository,
which makes it possible to build the go binary
as a "part" since the charmcraft.yaml is now in
a place where the build has access to the go code.

Similarly, the CI can be simplified and the spread
config too, because we can omit the logic to check
if the user has built the go binary before running
the tests.