experimentation: fix ubuntu-desktop-arm64 serial for Resolute Snapshot 2
Looking at the ISO tracker[1] for comparison, I noticed something
strange for Ubuntu Desktop arm64. The serial is from the 08 there.
Browsing in the folders, that's because all images are copied to the
next day before being overridden by a newer build, meaning latest serial
folder for Ubuntu Desktop is 20251209, but that folder still has an
arm64 build from 08.
That machinery is extremely confusing and error-prone 😢

[1]: https://iso.qa.ubuntu.com/qatracker/milestones/476/builds